backend/requirements.txt
# AI Chat Studio Backend Dependencies

# FastAPI核心
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# WebSocket支持
websockets==12.0

# 认证
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库 (可选)
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0  # PostgreSQL
aiomysql==0.2.0  # MySQL

# Redis缓存 (可选)
redis==5.0.1
aioredis==2.0.1

# AI模型集成
openai==1.3.5
anthropic==0.7.1

# HTTP客户端
httpx==0.25.2
aiohttp==3.9.1

# 工具库
python-dotenv==1.0.0
pyyaml==6.0.1

# 监控和日志
prometheus-client==0.19.0
loguru==0.7.2

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
